Move the implementation of ComponentInstancePreDestruction to a new dedicated class that follows the convention for application instance pre destructions. Delete DataWarningInstanciator and replace its delete method by the DataWarningInstancePreDestruction class that implements the ComponentInstancePreDestruction interface and its create method by the DataWarningInstancePostConstruction class tha implements the ComponentInstancePostConstruction interface. Delete all empty instanciator in the Silverpeas-Components subprojects (empty delete and empty create)
