+++ almanach/almanach-ejb/src/main/java/com/stratelia/webactiv/almanach/AlmanachInstancePreDestruction.java
@@ -1,53 +1,60 @@
 /**
- * Copyright (C) 2000 - 2013 Silverpeas
- *
+ * Copyright (C) 2000 - 2015 Silverpeas
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
- *
+ * <p>
  * As a special exception to the terms and conditions of version 3.0 of
  * the GPL, you may redistribute this Program in connection with Free/Libre
  * Open Source Software ("FLOSS") applications as described in Silverpeas's
  * FLOSS exception.  You should have received a copy of the text describing
  * the FLOSS exception, and it is also available here:
  * "http://www.silverpeas.org/docs/core/legal/floss_exception.html"
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.silverpeas.components.spaceMembers;
+package com.stratelia.webactiv.almanach;
 
-import com.silverpeas.admin.components.ComponentsInstanciatorIntf;
-import com.silverpeas.admin.components.InstanciationException;
-import com.stratelia.silverpeas.silvertrace.SilverTrace;
+import com.silverpeas.admin.components.ComponentInstancePreDestruction;
+import com.stratelia.webactiv.almanach.model.EventDAO;
+import org.silverpeas.util.DBUtil;
 
+import javax.inject.Named;
+import javax.transaction.Transactional;
 import java.sql.Connection;
 
 /**
- * @author
- * @version
+ * Before being deleted, remove all events in an almanach instance.
+ * @author mmoquillon
  */
-public class SpaceMembersInstanciator extends Object implements ComponentsInstanciatorIntf {
+@Named
+public class AlmanachInstancePreDestruction implements ComponentInstancePreDestruction {
 
-  /** Creates new SpaceMembersInstanciator */
-  public SpaceMembersInstanciator() {
+  private EventDAO eventDAO = new EventDAO();
+
+  /**
+   * Performs pre destruction tasks in the behalf of the specified Almanch instance.
+   * @param componentInstanceId the unique identifier of the Almanach instance.
+   */
+  @Transactional
+  @Override
+  public void preDestroy(final String componentInstanceId) {
+    try (Connection connection = DBUtil.openConnection()) {
+      getEventDAO().removeAllEvents(connection, componentInstanceId);
+    } catch (Exception e) {
+      throw new RuntimeException(e.getMessage(), e);
+    }
   }
 
-  @Override
-  public void create(Connection con, String spaceId, String componentId, String userId) throws
-      InstanciationException {
-
-  }
-
-  @Override
-  public void delete(Connection con, String spaceId, String componentId, String userId) throws
-      InstanciationException {
-
+  public EventDAO getEventDAO() {
+    return eventDAO;
   }
 }
